Simplification of the Wrapping work.
The "external world" should only manipulate `quickwit_search::SearchServiceClient` and `quickwit_search::SearchService`.

This new client has several functions:
- it embeds a SocketAddr, as was done for the WrappedSearchServiceClient
in the previous implementation
- it allows for calling directly a SearchService trait implementation,
which is useful both for mocking but also can be used to bypass the
transport when root == leaf node.
- it does the job of wrapping the tonic::Request and parsing the
SearchError automatically, simplifying the root code and the mocking.